+++ packages/patternfly-4/gatsby-theme-patternfly-org/gatsby-node.js
@@ -4,7 +4,6 @@ const { extractExamples } = require('./helpers/extractExamples');
 const { extractTableOfContents } = require('./helpers/extractTableOfContents');
 const { createHandlebars } = require('./helpers/createHandlebars');
 const { slugger } = require('./helpers/slugger');
-const webpack = require('webpack');
 
 // Add map PR-related environment variables to GraphQL
 exports.sourceNodes = ({ actions, createNodeId, createContentDigest }) => {
@@ -238,8 +237,8 @@ exports.createPages = ({ actions, graphql }, pluginOptions) => graphql(`
             actions.createPage({
               path: pagePath,
               component: source === 'core'
-                ? path.resolve(__dirname, './templates/fullscreenHtml.js')
-                : path.resolve(__dirname, './templates/fullscreenMdx.js'),
+                ? path.resolve(__dirname, './templates/fullscreenHTML.js')
+                : path.resolve(__dirname, './templates/fullscreenMDX.js'),
               context: {
                 // To exclude fullscreen pages from sitemap
                 isFullscreen: true,
@@ -288,21 +287,43 @@ exports.createSchemaCustomization = ({ actions }) => {
       pluginOptions: SitePluginOptions
     }
   `;
+
   actions.createTypes(sideNavTypeDefs);
 }
 
-exports.onCreateWebpackConfig = ({ actions, stage }) => {
-  if (stage === 'build-javascript') {
-    // Turn off source-maps because dist sizes are huge
-    actions.setWebpackConfig({
-      devtool: false
-    })
-  }
+exports.onCreateWebpackConfig = ({ actions, stage, getConfig }) => {
+  console.log(stage);
+  const config = getConfig();
+  
+  // Use caching for babel loader
+  const babelLoader = config.module.rules.find(rule => rule.test && rule.test.test('a.js'));
+  babelLoader.exclude = /node_modules(?!(\/gatsby-theme-patternfly-org))|@patternfly/;
+  const options = babelLoader.use[0].options;
+  options.cacheDirectory = '.cache/babel-loader';
+  options.cacheCompression = false;
+
   // Exclude CSS-in-JS styles included from React. They override
   // the patternfly.css styles which we would rather have.
-  actions.setWebpackConfig({
-    plugins: [
-      new webpack.NormalModuleReplacementPlugin(/react-styles\/css\/.*\.css/, path.resolve(__dirname, './empty.css'))
-    ]
+  const cssRules = config.module.rules.find(rule => rule.oneOf);
+  const cssRule = cssRules.oneOf.find(rule => rule.test.test('a.css'));
+  cssRule.exclude = /react-styles\/css\/.*\.css/;
+  config.module.rules.push({
+    test: /react-styles\/css\/.*\.css/,
+    use: 'null-loader'
   });
+
+  if (stage === 'build-javascript') {
+    // Turn off source-maps because dist sizes are huge
+    config.devtool = false;
+  } else if (stage === 'develop') {
+    // Speed up dev environment
+    config.devtool = 'cheap-source-map';
+    config.optimization = {
+      removeAvailableModules: false,
+      removeEmptyChunks: false,
+      splitChunks: false,
+    };
+  }
+
+  actions.replaceWebpackConfig(config);
 };