+++ packages/patternfly-4/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.50.9](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.8...patternfly-org-4@3.50.9) (2019-12-09)
+
+
+### Bug Fixes
+
+* **docs:** removed Blog masthead/footer links, pointed Get in touch page to Medium ([#1574](https://github.com/patternfly/patternfly-org/issues/1574)) ([a081fb9](https://github.com/patternfly/patternfly-org/commit/a081fb97d39af0f2b01b59c5554ade0d33c76c47))
+
+
+
+
+
+## [3.50.8](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.7...patternfly-org-4@3.50.8) (2019-12-06)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.50.7](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.6...patternfly-org-4@3.50.7) (2019-12-06)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.50.6](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.5...patternfly-org-4@3.50.6) (2019-12-04)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.50.5](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.4...patternfly-org-4@3.50.5) (2019-12-04)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.50.4](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.3...patternfly-org-4@3.50.4) (2019-12-03)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.50.3](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.2...patternfly-org-4@3.50.3) (2019-12-03)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.50.2](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.1...patternfly-org-4@3.50.2) (2019-12-03)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
 ## [3.50.1](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.50.0...patternfly-org-4@3.50.1) (2019-12-03)
 
 **Note:** Version bump only for package patternfly-org-4