+++ packages/patternfly-4/gatsby-theme-patternfly-org/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.10](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.9...gatsby-theme-patternfly-org@1.2.10) (2019-12-09)
+
+
+### Bug Fixes
+
+* **docs:** removed Blog masthead/footer links, pointed Get in touch page to Medium ([#1574](https://github.com/patternfly/patternfly-org/issues/1574)) ([a081fb9](https://github.com/patternfly/patternfly-org/commit/a081fb97d39af0f2b01b59c5554ade0d33c76c47))
+
+
+
+
+
+## [1.2.9](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.8...gatsby-theme-patternfly-org@1.2.9) (2019-12-06)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.8](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.7...gatsby-theme-patternfly-org@1.2.8) (2019-12-06)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.7](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.6...gatsby-theme-patternfly-org@1.2.7) (2019-12-04)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.6](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.5...gatsby-theme-patternfly-org@1.2.6) (2019-12-03)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.5](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.4...gatsby-theme-patternfly-org@1.2.5) (2019-12-03)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.4](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.3...gatsby-theme-patternfly-org@1.2.4) (2019-12-03)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
 ## [1.2.3](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.2...gatsby-theme-patternfly-org@1.2.3) (2019-12-03)
 
 **Note:** Version bump only for package gatsby-theme-patternfly-org